+++ src/settings.ts
@@ -2,6 +2,7 @@
  *
  * settings.ts: Settings and constants for homebridge-ratgdo.
  */
+import { RatgdoVariant } from "./ratgdo-types.js";
 
 // The platform the plugin creates.
 export const PLATFORM_NAME = "Ratgdo";
@@ -11,6 +12,12 @@ export const PLUGIN_NAME = "homebridge-ratgdo";
 
 // Interval, in seconds, to initiate mDNS discovery requests for new Ratgdo devices.
 export const RATGDO_AUTODISCOVERY_INTERVAL = 60;
+
+// mDNS TXT record project name associated with a Ratgdo device.
+export const RATGDO_AUTODISCOVERY_PROJECT_NAMES: string[] = [ RatgdoVariant.KONNECTED, RatgdoVariant.RATGDO ];
+
+// mDNS service types associated with a Ratgdo device.
+export const RATGDO_AUTODISCOVERY_TYPES = [ "esphomelib", "konnected" ];
 
 // Duration, in seconds, for a single heartbeat to ensure the Ratgdo doesn't autoreboot.
 export const RATGDO_HEARTBEAT_DURATION = 120;
@@ -33,3 +40,4 @@ export const RATGDO_OCCUPANCY_DURATION = 300;
 // Duration, in seconds, of our door state transition safety timer. This should be long enough that any reasonable state change message from Ratgdo would have had ample
 // time to be delivered, but short enough to provide that responsive feeling to an end user.
 export const RATGDO_TRANSITION_DURATION = 25;
+